requirements.txt
openpyxl==3.0.7
requests==2.25.1
Pillow==8.2.0
python_bcrypt==0.3.2
